bot/pyproject.toml
[project]
name = "governor"
version = "0.0.0"
description = "A Discord experience and leveling bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "discord-py>=2.4.0",
    "pillow>=10.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.pyright]
# reportAny = false
reportUnusedCallResult = false
